Completion of LogIn.html
The page should be adjustable and Resisable upto a certain window width, after which it would be classified as mobile, where ill create a different layout for later.

index.html is the landing page. At the moment there is nothing fancy, just two massive buttons that lead you to either client.html or LogIn.html.